Refactor augmented_diff to module
So we don't have to subprocess call a python function

- Fix crash when we try to write adiff to directory
  that doesn't exist by making the dir before writing
